Proper colorization mechanics, refactoring
