--- AccentApplicator/AccentControls.h
@@ -1,9 +1,0 @@
-#pragma once
-#include "framework.h"
-#include "AccentColorHelper.h"
-
-void Accent_Button(int* r, int* g, int* b, int* a);
-void Accent_CheckBox(int* r, int* g, int* b, int* a);
-void Accent_Tab(int* r, int* g, int* b, int* a);
-void Accent_Trackbar(int* r, int* g, int* b, int* a);
-void Accent_Spin(int* r, int* g, int* b, int* a);
