+++ AccentApplicator/BitmapHelper.cpp
@@ -1,0 +1,64 @@
+#include "BitmapHelper.h"
+
+#pragma region Pixel Color
+inline int PixClr(int val)
+{
+    return val & 0xFFFFFF;
+}
+
+inline int PixR(BYTE* pPixel)
+{
+    return PixClr(pPixel[2]);
+}
+inline int PixG(BYTE* pPixel)
+{
+    return PixClr(pPixel[1]);
+}
+inline int PixB(BYTE* pPixel)
+{
+    return PixClr(pPixel[0]);
+}
+inline int PixA(BYTE* pPixel)
+{
+    return PixClr(pPixel[3]);
+}
+#pragma endregion
+
+bool RecolorizeBitmap(HBITMAP hbm, BitmapHandler handler) {
+    BITMAP bm;
+    GetObject(hbm, sizeof(bm), &bm);
+
+    if (!hbm || bm.bmBitsPixel != 32)
+    {
+        return FALSE;
+    }
+
+    BYTE* pBits = new BYTE[bm.bmWidth * bm.bmHeight * 4];
+    GetBitmapBits(hbm, bm.bmWidth * bm.bmHeight * 4, pBits);
+
+    for (int y = 0; y < bm.bmHeight; y++)
+    {
+        BYTE* pPixel = (BYTE*)pBits + bm.bmWidth * 4 * y;
+
+        for (int x = 0; x < bm.bmWidth; x++)
+        {
+            int r = PixR(pPixel); // [2]
+            int g = PixG(pPixel); // [1]
+            int b = PixB(pPixel); // [0]
+            int a = PixA(pPixel); // [3]
+
+            handler(&r, &g, &b, &a);
+
+            pPixel[2] = r;
+            pPixel[1] = g;
+            pPixel[0] = b;
+            pPixel[3] = a;
+
+            pPixel += 4;
+        }
+    }
+
+    SetBitmapBits(hbm, bm.bmWidth * bm.bmHeight * 4, pBits);
+    delete[] pBits;
+    return TRUE;
+}
