+++ AccentApplicator/BitmapHelper.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "framework.h"
+
+typedef void (*BitmapHandler)(int* r, int* g, int* b, int* a);
+bool RecolorizeBitmap(HBITMAP hbm, BitmapHandler handler);
