+++ AccentApplicator/StyleModifier.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "framework.h"
+
+void ModifyStyles();
