+++ AccentApplicator/SysColorsModifier.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "framework.h"
+
+void ModifySysColors();
